fix styles fo dialog; use pipe in component; make orderBy more generic

+++ course-app/src/app/modules/main/container/course-list/course-list.component.ts
@@ -1,50 +1,50 @@
 import { Component, OnInit } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
+import 'rxjs/add/operator/filter';
 
 import { CourseListService } from './course-list.service';
 import { MatDialogService } from '../../../../core/dialogs/matDialog.service';
-import { ContainerCommunicationService } from '../container-communication.service';
+
+import { FilterSearchPipe } from './filter-search.pipe';
 
 import { Course } from '../../../../core/models/course.model';
 
 @Component({
   selector: 'app-course-list',
   templateUrl: './course-list.component.html',
-  styleUrls: ['./course-list.component.scss']
+  styleUrls: ['./course-list.component.scss'],
+  providers: [FilterSearchPipe]
 })
 export class CourseListComponent implements OnInit {
   courseList: Course[];
+  filteredList: Course[];
   filterField: string;
   isAsc: boolean;
-  sub: Subscription;
-  search: string;
 
   constructor(
     private courseListService: CourseListService,
     private matDialogService: MatDialogService,
-    private containerCommunicationService: ContainerCommunicationService
+    private filterSearchPipe: FilterSearchPipe,
   ) { }
 
   ngOnInit() {
-    this.sub = this.containerCommunicationService.channel$.subscribe(
-      data => this.search = data);
     this.filterField = 'startDate';
-    this.isAsc = true;
-    this.courseListService.getCoursesList().subscribe(courses => this.courseList = courses);
+    this.courseListService.getCoursesList().subscribe(courses => {
+      this.courseList = courses;
+      this.filteredList = courses;
+    });
   }
 
   deletedCourse(item: Course): void {
-    // TODO refactor
     this.matDialogService.show({
       title: 'Delete Action:',
       message: 'Do you really want to delete this course? ',
       icon: ''
     })
-    .subscribe(confirmed => {
-      if (confirmed) {
-        console.log(confirmed);
-        this.courseListService.deleteCourse(item);
-      }
-    });
+    .filter(confirmed => confirmed)
+    .subscribe(confirmed => this.courseListService.deleteCourse(item));
+  }
+
+  handleFilter(s: string): void {
+    this.filteredList = this.filterSearchPipe.transform(this.courseList, s);
   }
 }
